+++ NEWS.md
@@ -2,6 +2,7 @@
 ## IceSat2R 1.0.1
 
 * I've added the *'IceSat2R.R'* file with required dependencies
+* I've added an exception in the *time_specific_orbits()* function to remove internally processed sub-lists that do not have a 'description' (I expect that these sub-lists are not of type 'POINT')
 
 
 ## IceSat2R 1.0.0